AutoCR: excised all code for recently loaded sets
the new load-by-game-id mechanism should remove the need for this, and handling both workflows in a coherent way could be trouble. the two use cases for AutoCR are:

- CRs/devs reviewing an uploaded set, which is frozen
- devs/Jrs reviewing their own sets, which is changing

the former doesn't need to reload previously loaded sets, we can request them from the RA api again. the latter would want to review only the most up-to-date version, and wouldn't want to reload the previously loaded version